+++ src/features/mod.rs
@@ -11,3 +11,7 @@ pub(crate) mod resize_in_place;
 #[cfg(feature = "stats")]
 /// Allocation statistic gathering and reporting.
 pub mod stats;
+
+#[cfg(feature = "fallible_dealloc")]
+/// Fallible deallocation.
+pub mod fallible_dealloc;