+++ tests/stats.rs
@@ -1,3 +1,4 @@
+//#![allow(clippy::undocumented_unsafe_blocks)]
 use core::{
     alloc::Layout,
     sync::atomic::{AtomicUsize, Ordering},